refactor: extract workspace tab helpers
